samples: matter: template: Turn off buttons for 54L_ns
Turn off buttons for 54L_ns.

I don't know how the template/prj.conf file is able to enable this. I
observe that the DT uses gpio port 2 for button2. And that the nrfx
driver errors-out because it requires gpiote with this gpio port. But
the 54L HW does not support GPIOTE with gpio port 2.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>